add contigs to manual remap - televir
